pour some out in memory of ns-toggle-fullscreen
ELISP> emacs-version
"24.3.50.1"

emacs ending up with the LEAST USEFUL form of fullscreen support is just
uhh mindboggling at this point. ns-toggle-fullscreen was so vastly
superior to using osx's actual fullscreening.

and the release of 24.3 doesn't have the toggle-frame-fullscreen
function, and renders a gap at the top of the frame where osx's menu bar
would be. so you have to build from HEAD to get something that works at
all.

madness.