pi: use dnsmasq
I was trying to also use it to run a custom DHCP server that we could
tinker with for spanland, but the at&t router just gets in the way too
aggressively. Can't _really_ disable SLAAC/RDNSS on it, so best the pi
can do it manage IPv4, and then what's the point.

My android insists on accepting the DNS server from RDNSS and puts it
as the highest priority resolver, so the pi doesn't actually get used
from the phone. Very annoying. Erin's iphone seems to work fine and do
what I want.

Probably just gonna disable IPv6 on my phone until I finally give in
and buy a router to run behind the at&t one.