span: back to emacs-plus
It's just _perceptibly_ more performant. I suspect nativecomp wasn't
really kicking in on the nixpkgs copy, because this version has the
differently-annoying issue that the first N boots immediately froze
for 20+ minutes as compilation kicked in.

Had to `emacs -Q` then selectively eval parts of init.el until enough
had precompiled that things would actually respond.

Still want to move to the nixpkg, because there's no clear benefit to
using this copy that has to compile from source, but this is the path
of least resistance for now.