bin/openwrt-sysupgrade-image: download fresh openwrt images
Using this to manage packages instead of trying to install them at
runtime. rpt is stuck on a snapshot for the foreseeable future, where
packages become outdated and unreliable very quickly. And dewclaw
doesn't support apk yet, so I'll probably end up ripping all of that
code out in a fork anyhow.