openwrt images: +luci-app-package-manager
I don't use it to install anything, but it's occasionally nice to have a
UI to search for things when it's cumbersome to ssh, or I don't feel
like dealing with the opkg vs apk differences between my two routers.

Should maybe upgrade wrt to snapshot just to consolidate on apk.