tinkering with systemd-networkd
the goal/hope is that by using networkd, systemd will be less stupid
about managing the NFS mount on htpc, so that I don't occasionally
completely brick the machine because nixos restarts the network
with NFS still mounted and systemd loses its mind.

dunno if that will be true, and htpc is also on wifi which so far
seems like much more of a pain to set up with networkd.

maybe this works out maybe it doesn't