modules/httpd
this is almost certainly worse than just scattering caddy virtualHosts
entries everywhere, but I wanted to standardize the log mode and
didn't want to set it in a dozen places. so that became a nice
opportunity to mess around with types.submodule and the sorts of
hackery available there.

luckily everything i'm doing is dead simple. if i need to step outside
of the basics, this will immediately fall apart. lol.