tree-sitter overlay
I don't need to try to install anything as part of init.el, though
possibly on span I'll need to figure out how to add the right store
path to `treesit-extra-load-path`, since I'm not using a nix emacs
there.

not actually using the qmljs grammar rn, but I'm keeping it to
illustrate the injection point for adding extra grammars.