nixpkgs-terraform cachix
tho AFAICT it doesn't actually work because nothing I'd expect to be
available gets a cache hit.

https://docs.cachix.org/faq#why-is-nix-not-picking-up-on-any-of-the-pre-built-artifacts

So the terraform 1.12.2 package on desk is:

    nix-repl> pkgs.terraform-versions."1.12.2"
    evaluation warning: allowUnfree is enabled to build version 1.12.2
    «derivation /nix/store/p2bkjl45gh6wrx78dlwaanaz57xi2lcv-terraform-1.12.2.drv»

But that 404s when checking for the narinfo as their FAQ suggests:

    > curl -I https://nixpkgs-terraform.cachix.org/p2bkjl45gh6wrx78dlwaanaz57xi2lcv.narinfo
    HTTP/2 404
    date: Tue, 12 Aug 2025 13:17:12 GMT
    server: cloudflare

but hey maybe it ends up working before the next time I have to
upgrade terraform. and hey maybe I never have to upgrade terraform
again and every just adopts tofu as they should