desk: turn both displays on in one call
and disable the `layout-outputs` command for now. it's annoying to
have to run it myself with half the main display covered by what's
supposed to be the other monitor, but trying to confirm it's actually
the `--on` call(s) that are causing river to crash.

it's definitely _after_ swayidle resumes, so it's gotta be one of
these invocations.